document_chatbot/chatbot/requirements.txt

langchain
openai
faiss-cpu
python-dotenv
pypdf
tiktoken
streamlit
